refactor(buyer_leads): extract grupozap_buyer -> buyer transformation logic to grupozap_buyer module
