+++ apps/re/lib/leads/job_queue.ex
@@ -8,10 +8,7 @@ defmodule Re.Leads.Buyer.JobQueue do
   require Logger
 
   alias Re.{
-    Accounts.Users,
-    Leads.Buyer,
     Leads.GrupozapBuyer,
-    Listings,
     Repo
   }
 
@@ -20,7 +17,7 @@ defmodule Re.Leads.Buyer.JobQueue do
   def perform(%Multi{} = multi, %{"type" => "grupozap_buyer_lead", "uuid" => uuid}) do
     GrupozapBuyer
     |> Repo.get(uuid)
-    |> buyer_lead_changeset()
+    |> GrupozapBuyer.buyer_lead_changeset()
     |> insert_buyer_lead(multi)
     |> Repo.transaction()
   end
@@ -31,48 +28,5 @@ defmodule Re.Leads.Buyer.JobQueue do
     raise "Job not handled"
   end
 
-  defp buyer_lead_changeset(nil) do
-    Logger.warn("Leads.GrupozapBuyer not found")
-
-    raise("Leads.GrupozapBuyer not found")
-  end
-
-  defp buyer_lead_changeset(gzb) do
-    phone_number = concat_phone_number(gzb)
-
-    Buyer.changeset(%Buyer{}, %{
-      name: gzb.name,
-      email: gzb.email,
-      phone_number: phone_number,
-      origin: gzb.lead_origin,
-      user_uuid: extract_user_uuid(phone_number),
-      listing_uuid: extract_listing_uuid(gzb.client_listing_id)
-    })
-  end
-
-  defp concat_phone_number(%{ddd: nil, phone: nil}), do: "not informed"
-
-  defp concat_phone_number(%{ddd: _ddd, phone: nil}), do: "not informed"
-
-  defp concat_phone_number(%{ddd: nil, phone: phone}), do: "+55" <> phone
-
-  defp concat_phone_number(%{ddd: ddd, phone: phone}), do: "+55" <> ddd <> phone
-
   defp insert_buyer_lead(changeset, multi), do: Multi.insert(multi, :insert_buyer_lead, changeset)
-
-  defp extract_user_uuid("not informed"), do: nil
-
-  defp extract_user_uuid(phone_number) do
-    case Users.get_by_phone(phone_number) do
-      {:ok, user} -> user.uuid
-      _error -> nil
-    end
-  end
-
-  defp extract_listing_uuid(id) do
-    case Listings.get(id) do
-      {:ok, listing} -> listing.uuid
-      _error -> nil
-    end
-  end
 end
